scraper/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.12
google==3.0.0
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.73.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
oauthlib==3.2.2
packaging==25.0
playwright==1.52.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pyparsing==3.2.3
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
